LayoutTests/platform/ios-18/fast/forms/ios/form-control-refresh/search/textfield-appearance-background-expected.txt

Test that setting -webkit-appearance: textfield on a search input sets the background color to match the text input background color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textBackgroundColor is "rgb(255, 255, 255)"
FAIL searchBackgroundColor should be rgb(238, 238, 239). Was rgb(255, 255, 255).
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
